+++ packages/devextreme-cli/src/templates/vue-v2/sample-pages/home-page.vue
