+++ packages/devextreme-cli/src/templates/vue-v2/sample-pages/profile-page.vue
