+++ packages/devextreme-cli/src/templates/vue-v3/sample-pages/home-page.vue
