+++ packages/devextreme-cli/src/templates/vue-v3/sample-pages/profile-page.vue
