+++ packages/devextreme-cli/src/templates/vue-v3/sample-pages/tasks-page.vue
